// old/agir-work-planning-web/src/app/shared/models/findOptions/interventionFields.ts
export enum INTERVENTION_FIELDS {
  INTERVENTION_NAME = 'interventionName',
  PROGRAM_ID = 'programId',
  EXECUTOR_ID = 'executorId',
  ESTIMATE = 'estimate',
  PLANIFICATION_YEAR = 'planificationYear',
  DECESION_REQUIRED = 'decisionRequired',
  INTERVENTION_TYPE_ID = 'interventionTypeId',
  INTERVENTION_AREA = 'interventionArea',
  REQUESTOR_ID = 'requestorId',
  STREET_NAME = 'streetName',
  STREET_FROM = 'streetFrom',
  STREET_TO = 'streetTo',
  BOROUGH_ID = 'boroughId',
  STATUS = 'status',
  ASSETS = 'assets',
  ANNUALDISTRIBUTION_ANNUALPERIODS_YEAR = 'annualDistribution.annualPeriods.year',
  ANNUALDISTRIBUTION_ANNUALPERIODS_ANNUAL_LENGTH = 'annualDistribution.annualPeriods.annualLength',
  ANNUALDISTRIBUTION_ANNUALPERIODS_ANNUAL_ALLOWANCE = 'annualDistribution.annualPeriods.annualAllowance'
}